optimized and refactored code

+++ app/src/main/java/com/hvl/dat153/namequiz3/Utils/Utils.java
@@ -1,11 +1,9 @@
 package com.hvl.dat153.namequiz3.Utils;
 
-import android.app.Application;
 import android.content.Context;
 import android.content.ContextWrapper;
 import android.graphics.Bitmap;
 import android.graphics.BitmapFactory;
-import android.os.Environment;
 import android.util.Log;
 
 import java.io.File;
@@ -13,8 +11,6 @@ import java.io.FileOutputStream;
 import java.util.Random;
 
 public class Utils {
-
-    private Random random = new Random();
 
     private static final int size = 50;
 
@@ -43,7 +39,7 @@ public class Utils {
         FileOutputStream fos;
         try {
             fos = new FileOutputStream(mypath);
-            imageToSave.compress(Bitmap.CompressFormat.JPEG,100, fos);
+            imageToSave.compress(Bitmap.CompressFormat.JPEG, 100, fos);
             fos.close();
         } catch (Exception e) {
             Log.e("SAVE_IMAGE", e.getMessage());
@@ -51,18 +47,18 @@ public class Utils {
     }
 
     public static Bitmap getFile(String filename, Context context) {
-
         String path = context.getDir(Constants.PATH_TO_IMAGE_FOLDER, Context.MODE_PRIVATE).getAbsolutePath();
-        Bitmap recovered = BitmapFactory.decodeFile(path +"/"+ filename);
-        return recovered;
+        return BitmapFactory.decodeFile(path + "/" + filename);
     }
 
     public static void deleteFile(String filename, Context context) {
         File file = context.getDir(Constants.PATH_TO_IMAGE_FOLDER, Context.MODE_PRIVATE);
         File[] subfiles = file.listFiles();
 
-        for (File f : subfiles) {
-            if (f.getName().equals(filename)) f.delete();
+        if (subfiles != null) {
+            for (File f : subfiles) {
+                if (f.getName().equals(filename)) f.delete();
+            }
         }
     }
 }
